ogl_editor: linux/SDL code for everything
Finish the linux code, should be mostly coherent with win and mac.
Update the readme file too.

Just SDL basics for creating the window with opengl context, and passing
the mouse and keyboard inputs along. Menus are not displayed, but the
hotkeys are the same as in Windows. File dialogs use terminal input and
color dialog isn't implemented.

This is quite straightforward and platform-independent, so should be
pretty portable too.