skip track selection when no tracks available
Ignore arrow left/right actions if the "active track" has no group set,
which means that it's not a track. This fixes a crash that would happen
when attempting to access the current track's group, when the active
track is invalid (after starting the editor when nothing is connected,
or when the connection has no tracks).