Use explicit unsigned ints for EMGUI_COLOR32
Add "U" type suffixes in usage of the EMGUI_COLOR32 macro to pack colors
in uint32_t instead of having one cast for the whole expression in the
macro.

GCC 6.3.0 thinks that shifting signed values to the sign bit is a
non-constant expression and warns about it. It's probably just UB.