Fix build errors when compiling in Linux with CMakeLists.txt (#142)
* Ignore some warnings in external stb source

The stb libraries in emgui/src/External/ contain a bunch of silly code
that produces several warnings from gcc. Add -Wno-misleading-indentation
and -Wno-unused-but-set-variable for the offending files.

* Use explicit unsigned ints for EMGUI_COLOR32

Add "U" type suffixes in usage of the EMGUI_COLOR32 macro to pack colors
in uint32_t instead of having one cast for the whole expression in the
macro.

GCC 6.3.0 thinks that shifting signed values to the sign bit is a
non-constant expression and warns about it. It's probably just UB.

* Remove unused variables in Editor.c

Some local variables are assigned to but never read, so just delete
them.

* Fix indentation where gcc complains about it

Make -Werror=misleading-indentation pleased again by converting spaces
to tabs in a few places.

Should probably retab all tab-indented files as they contain a lot of
spaces but that's for another patch.