editor: add SET_KEY case to processCommands()
With this the demo can send SET_KEY to set a value in the editor
remotely from the demo context.

Imagine a demo with integrated developer tooling to do things like
place and orient cameras.  It would send the variables using SET_KEY
to persist them in the pattern data at the editor.

The way this is implemented currently, using doEdit(), the value is
echoed back to the demo just like if it were set interactively in
the editor.  Thus the value comes full circle and the demo may rely
on this exclusively for getting the new values realized in the normal
execution of its existing sync_update() and sync_get_val() machinery.