editor: add PAUSE case to processCommands()
With this the demo can send PAUSE to trigger a play/pause toggle
as if it had been done at the editor.

This is useful when in demo-centric editing mode, placing cameras
and such is only really practical while paused.

I've also made changes to librocket permitting SET_ROW changes to
continue arriving even when the demo is paused.  This way, the demo
may have fast-forward/rewind capabilities natively implemented, using
the scroll wheel wired to the current row variable for instance.

In combination with the ability to toggle PAUSE from the demo,
and with demo-originating SET_KEY support, the demo itself can be a
powerful component of the creative editing process.