Handle SET_ROW and PAUSE coming from the demo side (#149)
* editor: move processCommands() to bottom of listing

Preparatory change, so processCommands() can call more stuff in here
without requiring separate declarations.

* editor: split bare edit part from endEditing()

Moves core edit functionality into doEdit(), for reuse by future
SET_KEY case in processCommands().

* editor: add SET_KEY case to processCommands()

With this the demo can send SET_KEY to set a value in the editor
remotely from the demo context.

Imagine a demo with integrated developer tooling to do things like
place and orient cameras.  It would send the variables using SET_KEY
to persist them in the pattern data at the editor.

The way this is implemented currently, using doEdit(), the value is
echoed back to the demo just like if it were set interactively in
the editor.  Thus the value comes full circle and the demo may rely
on this exclusively for getting the new values realized in the normal
execution of its existing sync_update() and sync_get_val() machinery.

* editor: add PAUSE case to processCommands()

With this the demo can send PAUSE to trigger a play/pause toggle
as if it had been done at the editor.

This is useful when in demo-centric editing mode, placing cameras
and such is only really practical while paused.

I've also made changes to librocket permitting SET_ROW changes to
continue arriving even when the demo is paused.  This way, the demo
may have fast-forward/rewind capabilities natively implemented, using
the scroll wheel wired to the current row variable for instance.

In combination with the ability to toggle PAUSE from the demo,
and with demo-originating SET_KEY support, the demo itself can be a
powerful component of the creative editing process.