Implement support for multiple Rocket clients
This adds the ability for RocketEditor to accept multiple
simultaneous Rocket connections.

The connection which performs GET_TRACK is assumed to be the
Demo, which is a privileged client owning the map of track
names<->IDs.  If another client attempts to GET_TRACK or SET_KEY
than the current Demo client, that client is diconnected and an
error logged.  This single-demo limitation greatly simplifies
supporting multiple clients, as it skirts trying to turn the
track-ID mapping into more of a distributed mapping within the
existing Rocket protocol.

All clients however are permitted to send SET_ROW and PAUSE, with
these commands also being broadcast (with filtering) to all
clients for keeping them synchronized.

There's some ickiness to handling SET_ROW during playback, since
it's unlikely the clients are all *perfectly* aligned on their
current row.

When unpaused, incoming SET_ROW values are filtered to always
make forward progress.  This prevents the editor from jittering
backwards during playback due to minor discrepancies in the
clients current row.  The exception to this is when the delta is
at least a second worth of rows, which is seen as a deliberate
jump, regardless of direction.  The filtered SET_ROW values are
also prevented from propagating to other clients.

When paused however, any incoming SET_ROW is simply broadcast,
even the slightest difference.  As these are assumed to be
explicit precise movements during the creative track editing
process.  When broadcasting a SET_ROW, the originating client is
filtered from the recipients.  So clients don't need to do
anything WRT handling an echo.

The impetus for adding this feature is supporting use cases like
a tracker that has been modified to send PAUSE, and SET_ROW for
the pattern editor's cursor row up to RocketEditor.  This is
useful in coordinating visuals with music during the creative
process, with or without the demo showing visuals simultaneously
connected via Rocket.

I have an experimental branch where SchismTracker already does
this, so this has actually been somewhat successfully tested, on
Linux at least.

Note I've also added the TCP_NODELAY socket flag, which is a
change I've already had merged at the core Rocket project as
well.  When TCP_NODELAY is set everywhere in these multiclient
scenarios, it significantly helps keeping the jitter down.

The TCP_NODELAY addition may need some testing/changes for
handling non-Linux platforms, since that's all I've tested.